creating player entity and stats for movement and other posible stats for the player

+++ TinyBigIslandsOfJoris/TinyBigIslandsOfJoris/src/Game.hpp
@@ -13,8 +13,8 @@ public:
 
 	bool Init(const char* title, int xPos, int yPos, int width, int height, bool fullscreen);
 	void Ready();
-	void Update(float delta);
-	void PhysicsUpdate(float delta);
+	void Update(double delta);
+	void PhysicsUpdate(double delta);
 	void HandleEvents();
 	void Render();
 	void Finalize();
